+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: vvalet <vvalet@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/27 13:32:00 by vvalet            #+#    #+#             */
-/*   Updated: 2024/01/25 22:05:09 by vvalet           ###   ########.fr       */
+/*   Updated: 2024/01/25 22:06:42 by vvalet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	main(int argc, char **argv)
 	list_validity(argc, argv);
 	fill_stack(argc, argv, &a, &b);
 
-	int i;
+	size_t i;
 	// ft_push(&a, &b, 'b');
 	// ft_rrotate(&a, &b, 'a');
 	// ft_rotate(&a, &b, 'a');
